Refactor & change folder name of components
