+++ src/Components/users/style/Controls.css
@@ -18,28 +18,28 @@
   border-color: #bb0189e1;
 }
 .add-btn {
-  background-image: url(./../Icons/Plus\ Math.svg);
+  background-image: url(../../../assets/icons/Plus\ Math.svg);
   background-position: center;
   background-size: cover;
   background-size: 50%;
   background-repeat: no-repeat;
 }
 .edit-btn {
-  background-image: url(./../Icons/Mode\ edit.svg);
+  background-image: url(../../../assets/icons/Mode\ edit.svg);
   background-position: center;
   background-size: cover;
   background-size: 50%;
   background-repeat: no-repeat;
 }
 .list-btn {
-  background-image: url(./../Icons/Table\ chart.svg);
+  background-image: url(../../../assets/icons/Table\ chart.svg);
   background-position: center;
   background-size: cover;
   background-size: 50%;
   background-repeat: no-repeat;
 }
 .delete-btn {
-  background-image: url(./../Icons/Delete.svg);
+  background-image: url(../../../assets/icons/Delete.svg);
   background-position: center;
   background-size: cover;
   background-size: 50%;
